requirements.txt
PyQt6
sounddevice
soundfile
numpy
faster-whisper
google-generativeai
openai
anthropic
pyinstaller
